FEAT: add precompiler for netcdf
